bench-tps in a cargo test
